requirements.txt
ultralytics
mediapipe
pillow
colorama
